Remove Explorer.mostRecentActivity
This also moves the UI concerns of MostRecentActivity over to SplashScreen
